+++ model/parts/utils/utils.go
@@ -80,37 +80,8 @@ func GetPreferredChunkId() types.ChunkId {
 	return chunkId
 }
 
-func isThresholdFailed(firstNodeId types.NodeId, secondNodeId types.NodeId, graph *types.Graph, request types.Request) bool {
-	if config.GetThresholdEnabled() {
-		edgeDataFirst := graph.GetEdgeData(firstNodeId, secondNodeId)
-		p2pFirst := edgeDataFirst.A2B
-		edgeDataSecond := graph.GetEdgeData(secondNodeId, firstNodeId)
-		p2pSecond := edgeDataSecond.A2B
-
-		threshold := config.GetThreshold()
-		if config.IsAdjustableThreshold() {
-			threshold = edgeDataFirst.Threshold
-		}
-
-		peerPriceChunk := PeerPriceChunk(secondNodeId, request.ChunkId)
-
-		price := p2pFirst + peerPriceChunk
-		if config.GetReciprocityEnabled() {
-			price = p2pFirst - p2pSecond + peerPriceChunk
-		}
-		//fmt.Printf("price: %d = p2pFirst: %d - p2pSecond: %d + PeerPriceChunk: %d \n", price, p2pFirst, p2pSecond, peerPriceChunk)
-
-		if price > threshold {
-			if config.IsForgivenessEnabled() {
-				newP2pFirst, forgiven := CheckForgiveness(edgeDataFirst, firstNodeId, secondNodeId, graph, request)
-				if forgiven {
-					price = newP2pFirst - p2pSecond + peerPriceChunk
-				}
-			}
-		}
-		return price > threshold
-	}
-	return false
+func FindDistance(first types.NodeId, second types.ChunkId) int {
+	return config.GetBits() - general.BitLength(first.ToInt()^second.ToInt())
 }
 
 func getProximityChunk(firstNodeId types.NodeId, chunkId types.ChunkId) int {
@@ -144,16 +115,6 @@ func CreateDownloadersList(g *types.Graph) []types.NodeId {
 	//fmt.Println("Downloaders list create...!")
 	return downloadersList
 }
-
-// func CalculateStorageDepth(replication int) int {
-// 	depth := 0
-// 	n := config.GetNetworkSize()
-// 	for n/2 >= replication {
-// 		n = n / 2
-// 		depth++
-// 	}
-// 	return depth
-// }
 
 func CreateNodesList(g *types.Graph) []types.NodeId {
 	//fmt.Println("Creating nodes list...")
